changes in rooms structure
