--- src/main/java/com/campus/campus_hotel_artichaut_backend/entities/MediaType.java
@@ -1,7 +1,0 @@
-package com.campus.campus_hotel_artichaut_backend.entities;
-
-public enum MediaType {
-
-    IMAGE,
-    VIDEO,
-}
