--- src/main/java/com/campus/campus_hotel_artichaut_backend/entities/RoomType.java
@@ -1,8 +1,0 @@
-package com.campus.campus_hotel_artichaut_backend.entities;
-
-public enum RoomType {
-
-    STANDARD,
-    LUXE,
-    SUITE,
-}
