+++ src/main/java/com/campus/campus_hotel_artichaut_backend/model/entities/Duration.java
@@ -1,4 +1,4 @@
-package com.campus.campus_hotel_artichaut_backend.entities;
+package com.campus.campus_hotel_artichaut_backend.model.entities;
 
 public enum Duration {
     ONE_TIME,
