+++ src/main/java/com/campus/campus_hotel_artichaut_backend/model/entities/MediaType.java
@@ -1,0 +1,7 @@
+package com.campus.campus_hotel_artichaut_backend.model.entities;
+
+public enum MediaType {
+
+    IMAGE,
+    VIDEO,
+}
