+++ src/main/java/com/campus/campus_hotel_artichaut_backend/model/entities/Option.java
@@ -1,26 +1,32 @@
-package com.campus.campus_hotel_artichaut_backend.entities;
+package com.campus.campus_hotel_artichaut_backend.model.entities;
 
 import jakarta.persistence.Entity;
 import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
-import jakarta.validation.constraints.NotBlank;
+import jakarta.validation.constraints.NotNull;
 import jakarta.validation.constraints.PositiveOrZero;
 import lombok.Data;
 
-@Entity
 @Data
-public class Promotion {
+@Entity
+public class Option {
+
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private long id;
 
-    @NotBlank
     private String description;
 
     @PositiveOrZero
-    private float discount;
+    private int price;
 
-    @PositiveOrZero
-    private int minNumberOfNights;
+    private String name;
+
+    @NotNull
+    private Duration duration;
+
+    private boolean isCancellable;
+
+    private Reference reference;
 }
