+++ src/main/java/com/campus/campus_hotel_artichaut_backend/model/entities/Promotion.java
@@ -1,32 +1,26 @@
-package com.campus.campus_hotel_artichaut_backend.entities;
+package com.campus.campus_hotel_artichaut_backend.model.entities;
 
 import jakarta.persistence.Entity;
 import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
-import jakarta.validation.constraints.NotNull;
+import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.PositiveOrZero;
 import lombok.Data;
 
-@Data
 @Entity
-public class Option {
-
+@Data
+public class Promotion {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private long id;
 
+    @NotBlank
     private String description;
 
     @PositiveOrZero
-    private int price;
+    private float discount;
 
-    private String name;
-
-    @NotNull
-    private Duration duration;
-
-    private boolean isCancellable;
-
-    private Reference reference;
+    @PositiveOrZero
+    private int minNumberOfNights;
 }
