+++ src/main/java/com/campus/campus_hotel_artichaut_backend/repository/CustomerRepository.java
@@ -1,7 +1,6 @@
 package com.campus.campus_hotel_artichaut_backend.repository;
 
-import com.campus.campus_hotel_artichaut_backend.entities.Customer;
-import com.campus.campus_hotel_artichaut_backend.entities.RefreshToken;
+import com.campus.campus_hotel_artichaut_backend.model.entities.Customer;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
